refactor: organise components from CurriculumInfo folder
