+++ src/components/ArchivedComponents/CurriculumUnitDetails/CurriculumUnitDetails.tsx
@@ -5,7 +5,7 @@ import Box from "@/components/SharedComponents/Box";
 import { Heading, LI, P, OL } from "@/components/SharedComponents/Typography";
 import { TagFunctional } from "@/components/TagFunctional";
 import { Lesson } from "@/components/UnitModal/UnitModal";
-import { Thread } from "@/components/pages/CurriculumInfo/tabs/UnitsTab/UnitsTab";
+import { Thread } from "@/components/CurriculumComponents/UnitsTab/UnitsTab";
 import Accordion from "@/components/Accordion";
 
 export type CurriculumUnitDetailsProps = {
