+++ src/components/CurriculumComponents/OverviewTab/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./OverviewTab";
