+++ src/components/CurriculumComponents/helpers/curriculumMetadata.test.ts
