+++ src/components/CurriculumComponents/helpers/curriculumMetadata.ts
