+++ src/pages/teachers/curriculum/[subjectPhaseSlug]/[tab].tsx
@@ -10,8 +10,8 @@ import { useRouter } from "next/router";
 import CMSClient from "@/node-lib/cms";
 import { CurriculumOverviewSanityData } from "@/common-lib/cms-types";
 import CurriculumHeader from "@/components/CurriculumComponents/CurriculumHeader";
-import OverviewTab from "@/components/pages/CurriculumInfo/tabs/OverviewTab/OverviewTab";
-import UnitsTab from "@/components/pages/CurriculumInfo/tabs/UnitsTab/UnitsTab";
+import OverviewTab from "@/components/CurriculumComponents/OverviewTab";
+import UnitsTab from "@/components/CurriculumComponents/UnitsTab";
 import AppLayout from "@/components/AppLayout/AppLayout";
 import Box from "@/components/SharedComponents/Box";
 import curriculumApi, {
@@ -28,7 +28,7 @@ import { SubjectPhasePickerData } from "@/components/SharedComponents/SubjectPha
 import { fetchSubjectPhasePickerData } from "@/pages/teachers/curriculum/index";
 import getPageProps from "@/node-lib/getPageProps";
 import OakError from "@/errors/OakError";
-import { buildCurriculumMetadata } from "@/components/pages/CurriculumInfo/helpers/curriculumMetadata";
+import { buildCurriculumMetadata } from "@/components/CurriculumComponents/helpers/curriculumMetadata";
 
 export type CurriculumSelectionSlugs = {
   phaseSlug: string;
